+++ src/core/rect.c
@@ -1,6 +1,6 @@
 #include "rect.h"
 
-Rect rect_cut_top(Rect* r, size_t amount)
+Rect rect_cut_top(Rect* r, int32_t amount)
 {
     if (amount > r->h) {
         amount = r->h;
@@ -16,7 +16,7 @@ Rect rect_cut_top(Rect* r, size_t amount)
     return cut;
 }
 
-Rect rect_cut_bottom(Rect* r, size_t amount)
+Rect rect_cut_bottom(Rect* r, int32_t amount)
 {
     if (amount > r->h) {
         amount = r->h;
@@ -30,7 +30,7 @@ Rect rect_cut_bottom(Rect* r, size_t amount)
     };
 }
 
-Rect rect_cut_left(Rect* r, size_t amount)
+Rect rect_cut_left(Rect* r, int32_t amount)
 {
     if (amount > r->w) {
         amount = r->w;
@@ -46,7 +46,7 @@ Rect rect_cut_left(Rect* r, size_t amount)
     return cut;
 }
 
-Rect rect_cut_right(Rect* r, size_t amount)
+Rect rect_cut_right(Rect* r, int32_t amount)
 {
     if (amount > r->w) {
         amount = r->w;
@@ -60,12 +60,12 @@ Rect rect_cut_right(Rect* r, size_t amount)
     };
 }
 
-bool rect_contains(Rect r, size_t x, size_t y)
+bool rect_contains(Rect r, int32_t x, int32_t y)
 {
     return x >= r.x && x < r.x + r.w && y >= r.y && y < r.y + r.h;
 }
 
-Panel panel_cut_top(Rect* r, size_t height, size_t sep)
+Panel panel_cut_top(Rect* r, int32_t height, int32_t sep)
 {
     return (Panel) {
         .area = rect_cut_top(r, height),
@@ -73,7 +73,7 @@ Panel panel_cut_top(Rect* r, size_t height, size_t sep)
     };
 }
 
-Panel panel_cut_bottom(Rect* r, size_t height, size_t sep)
+Panel panel_cut_bottom(Rect* r, int32_t height, int32_t sep)
 {
     return (Panel) {
         .area = rect_cut_bottom(r, height),
@@ -81,7 +81,7 @@ Panel panel_cut_bottom(Rect* r, size_t height, size_t sep)
     };
 }
 
-Panel panel_cut_left(Rect* r, size_t width, size_t sep)
+Panel panel_cut_left(Rect* r, int32_t width, int32_t sep)
 {
     return (Panel) {
         .area = rect_cut_left(r, width),
@@ -89,7 +89,7 @@ Panel panel_cut_left(Rect* r, size_t width, size_t sep)
     };
 }
 
-Panel panel_cut_right(Rect* r, size_t width, size_t sep)
+Panel panel_cut_right(Rect* r, int32_t width, int32_t sep)
 {
     return (Panel) {
         .area = rect_cut_right(r, width),
